chore(tests): fixes
